+++ semantickernel-api/src/main/java/com/microsoft/semantickernel/contextvariables/ContextVariableTypeConverter.java
@@ -123,6 +123,22 @@ public class ContextVariableTypeConverter<T> {
     @Nullable
     @SuppressWarnings("unchecked")
     public <U> U toObject(ContextVariableTypes types, @Nullable Object t, Class<U> clazz) {
+        return toObject(types, t, clazz, true);
+    }
+
+    /**
+     * Use this converter to convert the object to the type of the context variable.
+     *
+     * @param <U>         the type to convert to
+     * @param t           the object to convert
+     * @param clazz       the class of the type to convert to
+     * @param logWarnings whether to log warnings
+     * @return the converted object
+     */
+    @Nullable
+    @SuppressWarnings("unchecked")
+    public <U> U toObject(ContextVariableTypes types, @Nullable Object t, Class<U> clazz,
+        boolean logWarnings) {
         if (t == null) {
             return null;
         }
@@ -145,8 +161,11 @@ public class ContextVariableTypeConverter<T> {
             return (U) converter.get().toObject((T) t);
         }
 
-        LOGGER.warn(SemanticKernelResources.getString("no.converter.found.for.to"), t.getClass(),
-            clazz);
+        if (logWarnings) {
+            LOGGER.warn(SemanticKernelResources.getString("no.converter.found.for.to"),
+                t.getClass(),
+                clazz);
+        }
         return null;
     }
 
@@ -173,7 +192,7 @@ public class ContextVariableTypeConverter<T> {
      * constructor.
      *
      * @param types the context variable types, if {@code null} the global types are used
-     * @param t the type to convert
+     * @param t     the type to convert
      * @return the prompt string
      */
     public String toPromptString(@Nullable ContextVariableTypes types, @Nullable T t) {